Consolidate code that interacts with device-lib
Consolidate code that handles forwarding the logs and holding the
context for interacting with device-lib in deviceLibWrapper.
