+++ src/Device/deviceLister.ts
@@ -6,63 +6,20 @@
 
 /* eslint-disable @typescript-eslint/ban-types */
 
-import { app } from '@electron/remote';
 import nrfDeviceLib, {
     Device as NrfdlDevice,
     DeviceTraits,
     HotplugEvent,
-    LogEvent,
-    setLogLevel,
-    setLogPattern,
 } from '@nordicsemiconductor/nrf-device-lib-js';
 import camelcaseKeys from 'camelcase-keys';
 import { Device, DeviceListing, Serialport } from 'pc-nrfconnect-shared';
 
-import { nrfdlVerboseLoggingEnabled } from '../Log/logSlice';
 import logger from '../logging';
 import { Devices } from '../state';
-import { getVerboseLoggingEnabled } from '../utils/persistentStore';
+import { getDeviceLibContext } from './deviceLibWrapper';
 import { devicesDetected } from './deviceSlice';
 
 const DEFAULT_DEVICE_WAIT_TIME = 3000;
-
-const deviceLibContext = nrfDeviceLib.createContext();
-nrfDeviceLib.setTimeoutConfig(deviceLibContext, { enumerateMs: 3 * 60 * 1000 });
-export const getDeviceLibContext = () => deviceLibContext;
-
-export const logNrfdlLogs =
-    (evt: LogEvent) => (_: Function, getState: Function) => {
-        if (app.isPackaged && !nrfdlVerboseLoggingEnabled(getState())) return;
-        switch (evt.level) {
-            case 'NRFDL_LOG_TRACE':
-                logger.verbose(evt.message);
-                break;
-            case 'NRFDL_LOG_DEBUG':
-                logger.debug(evt.message);
-                break;
-            case 'NRFDL_LOG_INFO':
-                logger.info(evt.message);
-                break;
-            case 'NRFDL_LOG_WARNING':
-                logger.warn(evt.message);
-                break;
-            case 'NRFDL_LOG_ERROR':
-                logger.error(evt.message);
-                break;
-            case 'NRFDL_LOG_CRITICAL':
-                logger.error(evt.message);
-                break;
-        }
-    };
-
-export const setDefaultNrfdlLogLevel = () =>
-    setLogLevel(getDeviceLibContext(), 'NRFDL_LOG_ERROR');
-
-if (getVerboseLoggingEnabled()) {
-    setLogLevel(getDeviceLibContext(), 'NRFDL_LOG_TRACE');
-} else setDefaultNrfdlLogLevel();
-
-setLogPattern(getDeviceLibContext(), '[%n][%l](%T.%e) %v');
 
 let hotplugTaskId: number;
 
@@ -159,7 +116,7 @@ export const startWatchingDevices =
  * @returns {undefined}
  */
 export const stopWatchingDevices = () => {
-    // Start here
+    // Not sure, if this guard clause is really needed
     if (getDeviceLibContext()) {
         try {
             nrfDeviceLib.stopHotplugEvents(hotplugTaskId);
