+++ src/index.ts
@@ -66,11 +66,11 @@ export { getAppSpecificStore as getPersistentStore } from './utils/persistentSto
 export { selectedDevice } from './Device/deviceSlice';
 export { deviceInfo } from './Device/deviceInfo/deviceInfo';
 export {
-    getDeviceLibContext,
     stopWatchingDevices,
     startWatchingDevices,
     waitForDevice,
 } from './Device/deviceLister';
+export { getDeviceLibContext } from './Device/deviceLibWrapper';
 export { default as sdfuOperations } from './Device/sdfuOperations';
 export { defaultInitPacket, HashType, FwType } from './Device/initPacket';
